+++ test/controllers/profile/niphl/RemoveNiphlControllerSpec.scala
@@ -19,7 +19,7 @@ package controllers.profile.niphl
 import base.SpecBase
 import base.TestConstants.{testEori, userAnswersId}
 import connectors.TraderProfileConnector
-import controllers.profile.niphl.routes._
+import controllers.profile.niphl.routes.*
 import forms.profile.niphl.RemoveNiphlFormProvider
 import models.{TraderProfile, UserAnswers}
 import navigation.{FakeProfileNavigator, ProfileNavigator}
@@ -31,7 +31,7 @@ import pages.profile.niphl.{HasNiphlUpdatePage, NiphlNumberUpdatePage, RemoveNip
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
-import play.api.test.Helpers._
+import play.api.test.Helpers.*
 import repositories.SessionRepository
 import views.html.profile.RemoveNiphlView
